check if the models with id from query exist

+++ PMCS/PMCS.API/Controllers/OwnerController.cs
@@ -1,6 +1,7 @@
 ﻿using AutoMapper;
 using FluentValidation;
 using Microsoft.AspNetCore.Mvc;
+using PMCS.API.Exceptions;
 using PMCS.API.Validators;
 using PMCS.API.ViewModels.Owner;
 using PMCS.DLL.Interfaces.Services;
@@ -34,6 +35,8 @@ namespace PMCS.API.Controllers
         [HttpGet("{id}")]
         public async Task<OwnerViewModel> GetById(int id, CancellationToken cancellationToken)
         {
+            if (!await IsModelExists(id, cancellationToken)) throw new ModelIsNotFoundException();
+
             return _mapper.Map<OwnerViewModel>(await _service.GetById(id, cancellationToken));
         }
 
@@ -50,6 +53,8 @@ namespace PMCS.API.Controllers
         [HttpDelete("{id}")]
         public async Task Delete(int id, CancellationToken cancellationToken)
         {
+            if (!await IsModelExists(id, cancellationToken)) throw new ModelIsNotFoundException();
+
             await _service.Delete(id, cancellationToken);
         }
 
@@ -61,7 +66,18 @@ namespace PMCS.API.Controllers
             var model = _mapper.Map<OwnerModel>(viewModel);
             model.Id = id;
 
+            if (!await IsModelExists(id, cancellationToken)) throw new ModelIsNotFoundException();
+
             return _mapper.Map<OwnerViewModel>(await _service.Update(model, cancellationToken));
+        }
+
+        private async Task<bool> IsModelExists(int id, CancellationToken cancellationToken)
+        {
+            var model = await _service.GetById(id, cancellationToken);
+
+            if (model == null) return false;
+
+            return true;
         }
 
     }
