+++ PMCS/PMCS.API/Controllers/PetController.cs
@@ -36,6 +36,8 @@ namespace PMCS.API.Controllers
         [HttpGet("{id}")]
         public async Task<PetViewModel> GetById(int id, CancellationToken cancellationToken)
         {
+            if (!await IsModelExists(id, cancellationToken)) throw new ModelIsNotFoundException();
+
             return _mapper.Map<PetViewModel>(await _service.GetById(id, cancellationToken));
         }
 
@@ -52,6 +54,8 @@ namespace PMCS.API.Controllers
         [HttpDelete("{id}")]
         public async Task Delete(int id, CancellationToken cancellationToken)
         {
+            if (!await IsModelExists(id, cancellationToken)) throw new ModelIsNotFoundException();
+
             await _service.Delete(id, cancellationToken);
         }
 
@@ -75,5 +79,15 @@ namespace PMCS.API.Controllers
 
             return pet.OwnerId;
         }
+
+        private async Task<bool> IsModelExists(int id, CancellationToken cancellationToken)
+        {
+            var model = await _service.GetById(id, cancellationToken);
+
+            if (model == null) return false;
+
+            return true;
+        }
+
     }
 }
